Finalized configurations page setup
